refactor(chat): streamline NewChatBootstrap and enhance state management
This commit refactors the NewChatBootstrap component to utilize the new `isNewChatBootstrapPreparing` function for determining bootstrap preparation status, improving clarity and maintainability. Additionally, it removes the dependency on the `usePrivy` hook, simplifying the authentication logic. The VercelChatProvider is also updated to pass the `isBootstrapPreparing` prop, ensuring consistent state management across components.
